KOGITO-5072: Isolate the dropdown items code.
