+++ packages/online-editor/src/editor/EditorToolbar.tsx
@@ -21,7 +21,6 @@ import {
   DropdownGroup,
   DropdownItem,
   DropdownPosition,
-  DropdownSeparator,
   DropdownToggle,
 } from "@patternfly/react-core/dist/js/components/Dropdown";
 import {
@@ -30,23 +29,18 @@ import {
   PageHeaderToolsGroup,
   PageHeaderToolsItem,
 } from "@patternfly/react-core/dist/js/components/Page";
-import { Text } from "@patternfly/react-core/dist/js/components/Text";
 import { TextInput } from "@patternfly/react-core/dist/js/components/TextInput";
 import { Title } from "@patternfly/react-core/dist/js/components/Title";
 import { Tooltip } from "@patternfly/react-core/dist/js/components/Tooltip";
 import { EllipsisVIcon } from "@patternfly/react-icons/dist/js/icons/ellipsis-v-icon";
-import { ExclamationTriangleIcon } from "@patternfly/react-icons/dist/js/icons/exclamation-triangle-icon";
 import * as React from "react";
 import { useCallback, useContext, useMemo, useState } from "react";
 import { useLocation } from "react-router";
 import { GlobalContext } from "../common/GlobalContext";
 import { useOnlineI18n } from "../common/i18n";
-import { DeploymentDropdownItem } from "./DmnDevSandbox/DeploymentDropdownItem";
-import { useDmnDevSandbox } from "./DmnDevSandbox/DmnDevSandboxContext";
-import { DmnDevSandboxInstanceStatus } from "./DmnDevSandbox/DmnDevSandboxInstanceStatus";
-import { useDmnRunner } from "./DmnRunner/DmnRunnerContext";
 import { useKieToolingExtendedServices } from "./KieToolingExtendedServices/KieToolingExtendedServicesContext";
 import { KieToolingExtendedServicesDropdown } from "./KieToolingExtendedServices/KieToolingExtendedServicesDropdown";
+import { useDropdownItems } from "./KieToolingExtendedServices/KieToolingExtendedServicesDropdownItems";
 import { KieToolingExtendedServicesStatus } from "./KieToolingExtendedServices/KieToolingExtendedServicesStatus";
 
 interface Props {
@@ -67,14 +61,13 @@ interface Props {
 export function EditorToolbar(props: Props) {
   const context = useContext(GlobalContext);
   const kieToolingExtendedServices = useKieToolingExtendedServices();
-  const dmnRunner = useDmnRunner();
-  const dmnDevSandbox = useDmnDevSandbox();
   const location = useLocation();
   const [fileName, setFileName] = useState(context.file.fileName);
   const [isShareMenuOpen, setShareMenuOpen] = useState(false);
   const [isViewKebabOpen, setViewKebabOpen] = useState(false);
   const [isKebabOpen, setKebabOpen] = useState(false);
   const { i18n } = useOnlineI18n();
+  const kieToolingExtendedServicesDropdownItems = useDropdownItems("sm");
 
   const logoProps = useMemo(() => {
     return { onClick: props.onClose };
@@ -218,155 +211,6 @@ export function EditorToolbar(props: Props) {
     [i18n, context, props.onSave, props.onDownload, props.onCopyContentToClipboard, props.onGistIt]
   );
 
-  const isKieToolingExtendedServicesRunning = useMemo(
-    () => kieToolingExtendedServices.status === KieToolingExtendedServicesStatus.RUNNING,
-    [kieToolingExtendedServices.status]
-  );
-
-  const onInstallKieToolingExtendedServices = useCallback(() => {
-    kieToolingExtendedServices.closeDmnTour();
-    kieToolingExtendedServices.setModalOpen(true);
-  }, [kieToolingExtendedServices]);
-
-  const onToggleDmnRunner = useCallback(() => {
-    kieToolingExtendedServices.closeDmnTour();
-    dmnRunner.setDrawerExpanded(!dmnRunner.isDrawerExpanded);
-  }, [dmnRunner, kieToolingExtendedServices]);
-
-  const onDevSandboxSetup = useCallback(() => {
-    kieToolingExtendedServices.closeDmnTour();
-    dmnDevSandbox.setConfigModalOpen(true);
-  }, [dmnDevSandbox, kieToolingExtendedServices]);
-
-  const onDevSandboxDeploy = useCallback(() => {
-    kieToolingExtendedServices.closeDmnTour();
-    dmnDevSandbox.setConfirmDeployModalOpen(true);
-  }, [dmnDevSandbox, kieToolingExtendedServices]);
-
-  const kieToolingExtendedServicesDropdownItems = useCallback(
-    (dropdownId: string) => {
-      const items = [
-        <DropdownItem
-          key={`kie-tooling-extended-services-dropdown-${dropdownId}-install`}
-          component={"button"}
-          onClick={onInstallKieToolingExtendedServices}
-          isDisabled={isKieToolingExtendedServicesRunning}
-          ouiaId="install-dropdown-button"
-        >
-          {isKieToolingExtendedServicesRunning
-            ? i18n.kieToolingExtendedServices.connectedOnPort(kieToolingExtendedServices.port)
-            : i18n.terms.install}
-        </DropdownItem>,
-        <DropdownGroup key={"dmn-runner-group"} label={i18n.names.dmnRunner}>
-          <Tooltip
-            data-testid={"dmn-runner-install-tooltip-for-open-toggle"}
-            content={i18n.kieToolingExtendedServices.dropdown.tooltip.installForDmnRunner}
-            trigger={!isKieToolingExtendedServicesRunning ? "mouseenter click" : ""}
-            position="left"
-          >
-            <DropdownItem
-              key={`kie-tooling-extended-services-dropdown-${dropdownId}-dmn-runner-toggle`}
-              component={"button"}
-              onClick={onToggleDmnRunner}
-              isDisabled={!isKieToolingExtendedServicesRunning}
-              ouiaId="toggle-dmn-runner-dropdown-button"
-              className="kogito--dmn-runner-button"
-            >
-              <Text>{dmnRunner.isDrawerExpanded ? i18n.terms.close : i18n.terms.open}</Text>
-            </DropdownItem>
-          </Tooltip>
-        </DropdownGroup>,
-        <DropdownGroup
-          key={"dmn-dev-sandbox-group"}
-          label={
-            <>
-              {i18n.names.dmnDevSandbox}
-              {isKieToolingExtendedServicesRunning &&
-                dmnDevSandbox.instanceStatus === DmnDevSandboxInstanceStatus.DISCONNECTED && (
-                  <Tooltip data-testid={"dmn-dev-sandbox-setup-tooltip"} content={i18n.dmnDevSandbox.common.setupFirst}>
-                    <ExclamationTriangleIcon className="pf-u-ml-sm" />
-                  </Tooltip>
-                )}
-            </>
-          }
-        >
-          <Tooltip
-            data-testid={"dmn-dev-sandbox-install-tooltip-for-setup"}
-            content={i18n.kieToolingExtendedServices.dropdown.tooltip.installForDmnDevSandbox}
-            trigger={!isKieToolingExtendedServicesRunning ? "mouseenter click" : ""}
-            position="left"
-          >
-            <DropdownItem
-              key={`kie-tooling-extended-services-dropdown-${dropdownId}-dmn-dev-sandbox-setup`}
-              component={"button"}
-              onClick={onDevSandboxSetup}
-              isDisabled={!isKieToolingExtendedServicesRunning}
-              ouiaId="setup-dmn-dev-sandbox-dropdown-button"
-            >
-              {i18n.terms.setup}
-            </DropdownItem>
-          </Tooltip>
-          <Tooltip
-            data-testid={"dmn-dev-sandbox-install-tooltip-for-deploy"}
-            content={i18n.kieToolingExtendedServices.dropdown.tooltip.installForDmnDevSandbox}
-            trigger={!isKieToolingExtendedServicesRunning ? "mouseenter click" : ""}
-            position="left"
-          >
-            <DropdownItem
-              key={`kie-tooling-extended-services-dropdown-${dropdownId}-dmn-dev-sandbox-deploy`}
-              component={"button"}
-              onClick={onDevSandboxDeploy}
-              isDisabled={
-                !isKieToolingExtendedServicesRunning ||
-                dmnDevSandbox.instanceStatus !== DmnDevSandboxInstanceStatus.CONNECTED
-              }
-              ouiaId="deploy-to-dmn-dev-sandbox-dropdown-button"
-            >
-              {i18n.terms.deploy}
-            </DropdownItem>
-          </Tooltip>
-        </DropdownGroup>,
-        <DropdownSeparator key={`kie-tooling-extended-services-dropdown-${dropdownId}-separator`} />,
-      ];
-
-      if (dmnDevSandbox.deployments.length === 0) {
-        items.push(
-          <DropdownItem
-            key={`kie-tooling-extended-services-dropdown-${dropdownId}-dmn-dev-sandbox-no-deployments`}
-            isDisabled
-          >
-            {i18n.dmnDevSandbox.dropdown.noDeployments}
-          </DropdownItem>
-        );
-      } else {
-        dmnDevSandbox.deployments
-          .sort((a, b) => b.creationTimestamp.getTime() - a.creationTimestamp.getTime())
-          .forEach((deployment, i) => {
-            items.push(
-              <DeploymentDropdownItem
-                id={i}
-                key={`kie-tooling-extended-services-dropdown-${dropdownId}-dmn-dev-sandbox-deployment_item_${i}`}
-                deployment={deployment}
-              />
-            );
-          });
-      }
-      return items;
-    },
-    [
-      dmnDevSandbox.deployments,
-      dmnDevSandbox.instanceStatus,
-      dmnRunner.isDrawerExpanded,
-      i18n,
-      isKieToolingExtendedServicesRunning,
-      kieToolingExtendedServices.port,
-      onDevSandboxDeploy,
-      onDevSandboxSetup,
-      onInstallKieToolingExtendedServices,
-      onToggleDmnRunner,
-    ]
-  );
-
   return !props.isPageFullscreen ? (
     <PageHeader
       logo={<Brand src={`images/${fileExtension}_kogito_logo.svg`} alt={`${fileExtension} kogito logo`} />}
@@ -385,7 +229,7 @@ export function EditorToolbar(props: Props) {
                   sm: "hidden",
                 }}
               >
-                <KieToolingExtendedServicesDropdown dropdownItems={kieToolingExtendedServicesDropdownItems} />
+                <KieToolingExtendedServicesDropdown />
               </PageHeaderToolsItem>
             </PageHeaderToolsGroup>
           )}
@@ -508,7 +352,7 @@ export function EditorToolbar(props: Props) {
                     key={"kie-tooling-extended-services-group"}
                     label={i18n.names.kieToolingExtendedServices}
                   >
-                    {...kieToolingExtendedServicesDropdownItems("sm")}
+                    {...kieToolingExtendedServicesDropdownItems()}
                   </DropdownGroup>,
                 ]}
                 position={DropdownPosition.right}
